Switch new config package from tools.gsf.config2 to tools.gsf.config, and delete legacy classes.
