--- gsf-core/src/main/java/tools/gsf/config/IcsObjectFactory.java
@@ -1,106 +1,0 @@
-/*
- * Copyright 2016 Function1. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tools.gsf.config;
-
-import COM.FutureTense.Interfaces.ICS;
-import tools.gsf.facade.assetapi.AssetAccessTemplate;
-import tools.gsf.facade.assetapi.asset.ScatteredAssetAccessTemplate;
-import tools.gsf.facade.assetapi.asset.TemplateAssetAccess;
-import tools.gsf.facade.mda.DefaultLocaleService;
-import tools.gsf.facade.mda.LocaleService;
-import tools.gsf.facade.search.SimpleSearchEngine;
-import tools.gsf.mapping.IcsMappingService;
-import tools.gsf.mapping.MappingService;
-import tools.gsf.navigation.AssetNode;
-import tools.gsf.navigation.LightweightSitePlanNavService;
-import tools.gsf.navigation.NavService;
-import tools.gsf.time.LoggerStopwatch;
-import tools.gsf.time.Stopwatch;
-
-/**
- * Factory implementation that works with a method naming convention to create
- * objects. Objects are created in a delegated factory method. The delegated
- * method is found by looking for a method that is prefixed by 'create' and then
- * the the simple name of the class (classname without package prefix).
- * <p>
- * For instance to create a object of class 'com.bar.Foo' if will look for a
- * method <tt>public Foo createFoo(ICS ics);</tt>. The method has to be public
- * and has to accept one argument of type ICS.
- * </p>
- *
- * @author Freddy Villalba
- * @author Tony Field
- * @since June 16, 2016
- */
-public class IcsObjectFactory extends BaseFactory {
-
-    /**
-     * @param ics Content Server context object
-     */
-    public IcsObjectFactory(ICS ics) {
-        this(ics, new Factory[0]);
-    }
-
-    /**
-     * @param ics   Content Server context object
-     * @param roots Factory object
-     */
-    public IcsObjectFactory(ICS ics, Factory... roots) {
-        super(ics, roots);
-    }
-
-    @ServiceProducer(cache = true)
-    public ScatteredAssetAccessTemplate createScatteredAssetAccessTemplate(final ICS ics) {
-        return new ScatteredAssetAccessTemplate(ics);
-    }
-
-    @ServiceProducer(cache = true)
-    public AssetAccessTemplate createAssetAccessTemplate(final ICS ics) {
-        return new AssetAccessTemplate(ics);
-    }
-
-    @ServiceProducer(cache = true)
-    public MappingService createMappingService(final ICS ics) {
-        return new IcsMappingService(ics);
-    }
-
-    @ServiceProducer(cache = true)
-    public LocaleService createLocaleService(final ICS ics) {
-        return new DefaultLocaleService(ics);
-    }
-
-    @ServiceProducer(cache = true)
-    public TemplateAssetAccess createTemplateAssetAccess(final ICS ics) {
-        return new TemplateAssetAccess(ics);
-    }
-
-    @ServiceProducer(cache = true)
-    public SimpleSearchEngine createSimpleSearchEngine(final ICS ics) {
-        return new SimpleSearchEngine("lucene");
-    }
-
-    @ServiceProducer(cache = true)
-    public NavService<AssetNode> createNavService(final ICS ics) {
-        TemplateAssetAccess dao = getObject("templateAssetAccess", TemplateAssetAccess.class);
-        return new LightweightSitePlanNavService(ics, dao);
-    }
-
-    @ServiceProducer(cache = false)
-    public Stopwatch createStopwatch(ICS ics) {
-        return LoggerStopwatch.getInstance();
-    }
-
-}
