+++ gsf-core/src/main/java/tools/gsf/config/ReflectionUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package tools.gsf.config2;
+package tools.gsf.config;
 
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
