+++ gsf-core/src/main/java/tools/gsf/config/ServiceProducer.java
@@ -25,7 +25,6 @@ import java.lang.annotation.Target;
  * indicates if this should be cached.
  *
  * @author Dolf Dijkstra
- *
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
