+++ gsf-core/src/main/java/tools/gsf/config/ServletContextLoader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package tools.gsf.config2;
+package tools.gsf.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,7 +26,7 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Set;
 
-import static tools.gsf.config2.ReflectionUtils.readConfigurationResource;
+import static tools.gsf.config.ReflectionUtils.readConfigurationResource;
 
 /**
  * ServletContextListener that loads and configures the FactoryProducer for this
