+++ gsf-core/src/main/java/tools/gsf/config/inject/InjectForRequest.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package tools.gsf.config.inject;
 
 import java.lang.annotation.Documented;
