--- gsf-core/src/main/java/tools/gsf/config2/ServletContextBackedFactory.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2016 Function1. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tools.gsf.config2;
-
-import tools.gsf.time.LoggerStopwatch;
-import tools.gsf.time.Stopwatch;
-
-import javax.servlet.ServletContext;
-
-/**
- * @author Tony Field
- * @since 2016-08-05
- */
-public class ServletContextBackedFactory extends AbstractDelegatingFactory<ServletContext> {
-
-    private final ServletContext servletContext;
-
-    public ServletContextBackedFactory(ServletContext servletContext, Factory delegate) {
-        super(servletContext, delegate);
-        this.servletContext = servletContext;
-    }
-
-    @ServiceProducer
-    public Stopwatch newStopwatch() {
-        return LoggerStopwatch.getInstance();
-    }
-}
