--- gsf-core/src/main/java/tools/gsf/config2/inject/InjectForRequest.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2011 FatWire Corporation. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package tools.gsf.config2.inject;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Annotation used to indicate that the field should have its value injected by
- * the Injector.
- *
- * @author Dolf Dijkstra
- * @since 2011-03-24
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD, ElementType.METHOD})
-@Documented
-public @interface InjectForRequest {
-
-    /**
-     * @return the name of the object in the Factory.
-     */
-    String value() default "";
-
-}
