--- gsf-core/src/main/java/tools/gsf/config2/inject/Injector.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2016 Function1. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tools.gsf.config2.inject;
-
-/**
- * Implementations inject services into the dependent object.
- * <p>
- * See <a href="http://en.wikipedia.org/wiki/Dependency_injection">Dependency Injection</a>
- * for more information on Dependency Injection.
- *
- * @author Dolf.Dijkstra
- *
- */
-public interface Injector {
-
-    void inject(Object dependent);
-
-}
